requirements.txt
django==5.0.6 
djangorestframework==3.15.1
channels==4.1.0
djangorestframework-simplejwt==5.3.1
daphne==4.1.2


